Instantiate getElementsByClassName.js
Working towards a solution for js file.
First on the agenda is to research to
know what getElementsByClassName does.

After will be to write some sort of
outline of how the function should work.
To note: function must utilize the
methods listed in the outline page of
the recursion precourse.