Resolve object key/value bug
Utilizing my first initial solution
receive passing test from test suite
by applying the stringifyJSON function
upon the value of each property in an
object.

The reason being, was that I could not
get a passing solution by simply having
the value (if it was a string) in quotation
marks.  Rather, it must pass the function
recursively, especially if the value of the
property is again <strong> another array or
object </strong>.