Render functional getElementsByClassName
Pass the mocha tests for the function, and
corrected the tests I had created to check
for mistakes.  Forgot that data by
reference renders false when compared to
other data since they point to data that
lies within the memory and not to
primitive types.

Now to refactor for empathy