Finalize getElementsByClassName.js
Finished function, and refactored for
reasonable enough empathy.

The function utilizes two helper functions:
(1) currentNodeHassClass and (2) renderCollection

getElementsByClassName broken down does only
one thing.  The function collects all the nodes
that have the specified class. However, the
function must traverse the tree data structure
of the document's body which is a node comprised
of several child nodes. Meaning that iterating
through all the possible items is no longer
linear.  The function needs to be flexible enough
to traverse not only the current nodes class list
but the class lists of its children, and the child
nodes must also be traversed if they too have
childNodes.

That is why getElementsByClassName requires the
renderCollection function which comprises of two
parts.
(1) Check the current nodes classlist
(2) Check to see if there are any child nodes
In the case where there are child nodes the
function repeats the two instructions recursively
unless the node no longer has any more childNodes.
In the case that the node does not have any more
children, then the collection is rendered.